% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempFunctions.R
\name{celsToFah}
\alias{celsToFah}
\title{Celsius to Fahrenheit Conversion}
\usage{
celsToFah(c)
}
\arguments{
\item{c}{A numeric Celsius value to be converted to Fahrenheit}
}
\value{
Fahrenheit conversion
}
\description{
Takes in a numeric Celsius value and converts it to Fahrenheit
}
