% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempFunctions.R
\name{fahToCels}
\alias{fahToCels}
\title{Fahrenheit to Celsius Conversion}
\usage{
fahToCels(f)
}
\arguments{
\item{f}{A numeric Fahrenheit value to be converted to Celsius}
}
\value{
Celsius conversion
}
\description{
Takes in a numeric Fahrenheit value and converts it to Celsius
}
